test: Split CBC-solver test into separate file
